Fix parametric gear with modern Fusion versions.
It appears that Fusion has added control handles to bevel curves that broke the
gear tooth profile when changing number of teeth or other parameters. I deleted
the extra control handles and made other small changes to fix the model.

Also changed from a filet to a chamfer on the edges of the gear teeth, since
chamfers work much better for 3D printing due to a consistent overhang angle.

Removed the version number from the .f3d file name, since version history isn't
exported and it's not really relevant to users of the model.